Disallow db pull in production

+++ lib/tasks/db.rake
@@ -3,6 +3,8 @@
 namespace :db do
   desc 'Replaces the local database with a copy downloaded from production'
   task pull: :environment do
+    raise "Do not use in production!" if Rails.env == "production"
+
     sh 'rm -f latest.dump'
     sh 'heroku pg:backups:capture'
     sh 'heroku pg:backups:download'
